requirements.txt
# Manage environment variables with python-dotenv
python-dotenv==0.20.0  # Compatible with Python 3.10

# Formatters & Linters
black==22.8.0           # Compatible with Python 3.10
pylint==2.15.0          # Compatible with Python 3.10

# Testing
pytest==7.1.2           # Fully compatible
fire==0.5.0             # Latest version is compatible

# CLI & Utilities
click==8.1.3            # Compatible with Python 3.10
ipython==8.4.0          # Python 3.10 support starts from 8.0

# Streamlit
streamlit==1.13.0       # Known to work well with Python 3.10
altair==4.2.2

# OpenAI
openai>=1.12.0          # 1.12.0+ is tested for Python 3.10

# Whisper
git+https://github.com/openai/whisper.git